chore(refactor): refactor self-check widget
