+++ lib/src/features/scan/presentation/widgets/connected_device.dart
@@ -12,42 +12,53 @@ class ConnectedDevice extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      width: context.width * 0.35,
+      width: context.width * 0.3,
       height: 180,
       decoration: BoxDecoration(
-        color: context.scheme.primary.withOpacity(.34),
+        color: context.scheme.primary,
         borderRadius: const BorderRadius.only(
           topRight: Radius.circular(22),
           bottomRight: Radius.circular(22),
         ),
       ),
-      child: Center(
-        child: Column(
-          children: [
-            8.vGap,
-            Text(
-              'Intellibra',
-              style: context.bodyLg.copyWith(
-                color: context.scheme.onPrimary,
+      child: Container(
+        width: context.width * 0.35,
+        height: 180,
+        decoration: BoxDecoration(
+          color: context.scheme.primary.withOpacity(.34),
+          borderRadius: const BorderRadius.only(
+            topRight: Radius.circular(22),
+            bottomRight: Radius.circular(22),
+          ),
+        ),
+        child: Center(
+          child: Column(
+            children: [
+              8.vGap,
+              Text(
+                'Intellibra',
+                style: context.bodyLg.copyWith(
+                  color: context.scheme.onPrimary,
+                ),
               ),
-            ),
-            ShadowOverlay(
-              shadowHeight: 60,
-              shadowWidth: context.width * 0.35,
-              shadowColor: context.scheme.primary.withOpacity(0.9),
-              child: Image.asset(
-                Assets.assetsIconsIntellibra,
+              ShadowOverlay(
+                shadowHeight: 60,
+                shadowWidth: context.width * 0.35,
+                shadowColor: context.scheme.primary.withOpacity(0.9),
+                child: Image.asset(
+                  Assets.assetsIconsIntellibra,
+                ),
               ),
-            ),
-            8.vGap,
-            Text(
-              'CE12XFMZ',
-              style: context.bodyMd.copyWith(
-                color: context.scheme.onPrimary,
-                fontWeight: FontWeight.bold,
+              8.vGap,
+              Text(
+                'CE12XFMZ',
+                style: context.bodyMd.copyWith(
+                  color: context.scheme.onPrimary,
+                  fontWeight: FontWeight.bold,
+                ),
               ),
-            ),
-          ],
+            ],
+          ),
         ),
       ),
     );
